medical-dictation/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
faster-whisper>=1.1.0
python-dotenv>=1.0.1
numpy>=1.26.2
pydantic>=2.8.0
